Move layout components, make them more generic
