+++ src/applications/vaos/new-appointment/components/FormLayout.jsx
@@ -3,7 +3,7 @@ import { Link } from 'react-router-dom';
 import Breadcrumbs from '../../components/Breadcrumbs';
 import NeedHelp from '../../components/NeedHelp';
 
-export default function NewAppointmentLayout({ children, isReviewPage }) {
+export default function FormLayout({ children, isReviewPage }) {
   return (
     <div className="vads-l-grid-container vads-u-padding-x--2p5 large-screen:vads-u-padding-x--0 vads-u-padding-bottom--2">
       <Breadcrumbs>
